net: Extend flow info with IP addresses from ICMP data
When selecting the next hop, in case of ECMP, we would like to treat
ICMP error messages in a special way, that is route them together with
the flow they belong to. To do that we need to know the source and
destination address of the original IP packet that triggered the
error (see ip_multipath_icmp_hash).

Make room for this info in flowi4, flowi6 objects which we pass down the
call chain of the routing subsystem, past the point where we still have
access to the SKB.

Signed-off-by: Jakub Sitnicki <jkbs@redhat.com>