PoC: ipv4: Don't promote secondaries when flushing addresses of device
Promoting secondary addresses on address removal makes flushing
addresses from devices with 1000's of them slow. This is because we
cannot take down the secondary addresses when we are removing the
primary one (which would make it faster).

The userspace, when performing a flush, will in the end remove all the
addresses, regardless of secondary address promotion taking
place. Unfortunately, the kernel doesn't actually know when the
userspace is asking for a removal of just a particular address and when
it is just a removal of one of the addresses due to a flush operation in
progress.

To help with this case, introduce a IFA_F_FLUSH flag that can be used by
userspace to signal that a removal operation is being done because of an
address flush. When the flag is set, don't bother with secondary address
promotion as we expect that secondary addresses will be removed soon as
well.

It's also worth noting that promote_secondaries sysctl param is enabled
by default since systemd 216, thus making it the new "normal".

Signed-off-by: Jakub Sitnicki <jkbs@redhat.com>