Make TIME-WAIT reuse delay deterministic and configurable
This is an iteration on an effort to make the TIME-WAIT reuse delay
shorter, which we have recently presented about at Plumbers [1].

I have addressed the RFCv1 feedback and reworked the changes so that the
scope is limited to just the TIME-WAIT reuse code and does not touch the
PAWS implementation. Please see patch 1 description for the discussion of
this approach.

The end result is that with this patch set the TS.Recent last update
timestamp is interpreted as a millisecond value only in the true TIME-WAIT
state.

I feel that switching the TS.Recent last update timestamp to milliseconds
everywhere in one go would needlessly bundle up the risk of causing
regressions in both TIME-WAIT reuse and PAWS detection code.

Unless there is a strong guidance from the maintainers to do it all at
once, I'd sleep better if we could do it in steps and attack just the
TIME-WAIT reuse first.

This patch set is accompanied by a set of packetdrill tests for both
slow (after 1 sec) and fast (after 1 msec) TIME-WAIT reuse covering happy
and failure scenarios. They can be reviewed at [2]. If the proposed changes
make it into the kernel, I will to submit a PR to the packetdrill repo.

I also plan on adding coverage for PAWS old duplicate detection, as I don't
think we have any in the packetdrill repo. They would be part of the
follow-up changes where we would have to touch the PAWS code to use
milliseconds for TS.Recent last update timestamp everywhere.

We will be rolling these changes out internally to a limited set of
production machines to catch any potential regressions before posting
another iteration. We will report on any findings.

Credit is due for Adrien Vasseur and Lee Valentine for the initial report
on how use of IP_LOCAL_PORT_RANGE when TIME-WAIT reuse delay it up to 1
second long can lead to port exhaustion under connection pressure.

Goes without saying - we are looking for your feedback, while we test this.

Thanks,
-jkbs

[1] https://lpc.events/event/18/contributions/1962/ 
[2] https://github.com/jsitnicki/packetdrill/tree/tw-reuse-tests/rfc2/gtests/net/tcp/ts_recent/tw_reuse

# Describe the purpose of this series. The information you put here
# will be used by the project maintainer to make a decision whether
# your patches should be reviewed, and in what priority order. Please be
# very detailed and link to any relevant discussions or sites that the
# maintainer can review to better understand your proposed changes. If you
# only have a single patch in your series, the contents of the cover
# letter will be appended to the "under-the-cut" portion of the patch.

# Lines starting with # will be removed from the cover letter. You can
# use them to add notes or reminders to yourself. If you want to use
# markdown headers in your cover letter, start the line with ">#".

# You can add trailers to the cover letter. Any email addresses found in
# these trailers will be added to the addresses specified/generated
# during the b4 send stage. You can also run "b4 prep --auto-to-cc" to
# auto-populate the To: and Cc: trailers based on the code being
# modified.

To: netdev@vger.kernel.org
Cc: Eric Dumazet <edumazet@google.com>
Cc: Jason Xing <kerneljasonxing@gmail.com>
Cc: Adrien Vasseur <avasseur@cloudflare.com>
Cc: Lee Valentine <lvalentine@cloudflare.com>
Cc: kernel-team@cloudflare.com
Signed-off-by: Jakub Sitnicki <jakub@cloudflare.com>

---
Changes in RFCv2:
- Make TIME-WAIT reuse configurable through a per-netns sysctl.
- Account for timestamp rounding so delay is not shorter than set value.
- Use tcp_mstamp when we know it is fresh due to receiving a segment.
- Link to RFCv1: https://lore.kernel.org/r/20240819-jakub-krn-909-poc-msec-tw-tstamp-v1-1-6567b5006fbe@cloudflare.com

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 2,
    "change-id": "20240815-jakub-krn-909-poc-msec-tw-tstamp-05aad5634a74",
    "prefixes": [
      "RFC net-next"
    ],
    "history": {
      "v1": [
        "20240815-jakub-krn-909-poc-msec-tw-tstamp-v1-1-a9598c1309fc@cloudflare.com",
        "20240819-jakub-krn-909-poc-msec-tw-tstamp-v1-1-6567b5006fbe@cloudflare.com"
      ]
    }
  }
}