trait: XDP support
Store the trait KV store in the xdp_buff headroom, right after the
xdp_frame.

Support for traits is gated on support for metadata: to propagate the
traits to the skb layer, drivers will have to explicitly signal support
like the existing skb_metadata_set() call.

This also makes bounds checks simpler - we can rely on xdp->data_meta
always being a valid hard_end for traits.

This assumes there is enough headroom to store the traits header. That
avoids having to check in the hot-path, where all we could do anyways is
BUG_ON().

Signed-off-by: Arthur Fabre <afabre@cloudflare.com>