trait: Propagate presence of traits to sk_buff
Steal a flag from skb_shared_info to store whether or not an skb stores
traits in it's headroom.
The metadata length is already stored there, so setting it shouldn't be
slow.
skb_shared_info also s naturally persistent across clones, which makes
sense for the traits flag.

But it does raise the question: what should / can we for cloned SKBs?
Reading traits isn't a problem, but setting a trait would set it for all
the clones...

Have skb_metadata_set() set the flag, so that all drivers that currently
support metadata automatically support traits too.

Signed-off-by: Arthur Fabre <afabre@cloudflare.com>